dockerfiles/whisper-server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai-whisper>=20231117
torch>=2.1.0
